Extract temporary key template and copy
Additional warnings, etc. need to be added.
Extracting copy and separating everything related to temp keys into its
own file makes it easier to modify later.
