+++ src/frontend/src/flows/manage/authenticatorsSection.ts
@@ -15,7 +15,9 @@ const MAX_AUTHENTICATORS = 8;
 export type DedupAuthenticator = Authenticator & { dupCount?: number };
 
 // Deduplicate devices with same (duplicated) aliases
-const dedupLabels = (authenticators: Authenticator[]): DedupAuthenticator[] => {
+export const dedupLabels = (
+  authenticators: Authenticator[]
+): DedupAuthenticator[] => {
   return authenticators.reduce<Authenticator[]>((acc, authenticator) => {
     const _authenticator: DedupAuthenticator = { ...authenticator };
     const sameName = acc.filter((a) => a.alias === _authenticator.alias);
@@ -107,32 +109,6 @@ export const authenticatorsSection = ({
 
         </div>
     </aside>`;
-};
-
-export const tempKeysSection = ({
-  authenticators: authenticators_,
-}: {
-  authenticators: Authenticator[];
-}): TemplateResult => {
-  const authenticators = dedupLabels(authenticators_);
-
-  return html` <aside class="l-stack c-card c-card--narrow">
-    <div class="t-title t-title--complications">
-      <h2 class="t-title">Temporary Keys</h2>
-    </div>
-
-    <p style="max-width: 30rem;" class="t-paragraph t-lead">
-      These keys are stored in browser cache, clearing your browser storage will
-      erase them.
-    </p>
-    <div class="c-action-list">
-      <ul>
-        ${authenticators.map((authenticator, index) =>
-          authenticatorItem({ authenticator, index })
-        )}
-      </ul>
-    </div>
-  </aside>`;
 };
 
 export const authenticatorItem = ({
