+++ src/frontend/src/flows/manage/tempKeys.json
@@ -1,0 +1,6 @@
+{
+  "en": {
+    "temporary_key": "Temporary Keys",
+    "key_stored_in_browser": "These keys are stored in browser cache, clearing your browser storage will erase them."
+  }
+}
